Update SQS Controller deps, generator and re-generate controller (#7)
Issue #, if available: [https://github.com/aws-controllers-k8s/community/issues/205]

Description of changes:

As per the discussion on issue #205, this PR does the following

- Updates the controller to Go 1.17
- Updates the runtime package to v0.18.4
- Re-generates the SQS controller with an updated generator.yaml
- Provides a basic e2e smoke test, where the queue resource is created and deleted.

There are still known issues with the controller, such as
- If the spec doesnt includes any attributes, the SQS API call for Create will fail
- Updating the resource currently doesn't work, as the sdkRead function doesn't assign the attributes to properties that the delta functions can use

IMO, the two known issues may be addressed in the generator project, which is where I anticipate heading next. It's my first PR in this project, absolutely keen for feedback/improvements. Hope this is helpful after all.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.