fix: ignore updates if attributes are empty
This change is required for users migrating from an older sqs controller
(pre sdk v2 migration) to the latest one. In aws sdk go v1, the sdk
update was able to send empty attributes, and no error was returned.
When we migrated to v2, empty attributes when calling SetQueueAttributes
was no longer an option. We were able to mitigate this issue by late
initializing the defaulted fields when creating a queue to avoid
exposing them in the delta. This mitigation did not handle cases in
which users migrate to the latest version, as the resource is not late
initialized if update fails. With this change, we will avoid calling
update if there are no attributes defined by the user