fix: ignore attributes update empty (#96)
Issue [#2247](https://github.com/aws-controllers-k8s/community/issues/2247)

Description of changes:
This change is required for users migrating from an older sqs controller
(pre sdk v2 migration) to the latest one. In aws sdk go v1, the sdk
update was able to send empty attributes, and no error was returned.
When we migrated to v2, empty attributes when calling SetQueueAttributes
was no longer an option. We were able to mitigate this issue by late
initializing the defaulted fields when creating a queue to avoid
exposing them in the delta. This mitigation did not handle cases in
which users migrate to the latest version, as the resource is not late
initialized if update fails. With this change, we will avoid calling
update if there are no attributes defined by the user

In sdk v1, the attributes input was a `*map[string]string`, so if it was
 not set, it would just be a No-Op

In sdk v2, attributes is `map[string]string`, so if it's not set, it would 
default to an empty map, which started causing issues

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.